Adding scripts for getting data into dev_appserver.
These changes do not change the way the application runs on GAE but are
necessary for locally testing operations that operate on data.

server/download_dev_data.sh:
There is a documented procedure to download data from the datastore
hosted on GAE to a local SQLite3 file using the appcfg.py utility, but
it is a little cumbersome so why not sutomate it.  This does not pruduce
data that the dev_appserver can use so we need to "upload" the data we
just downloaded into the dev_appserver.

server/upload_local_data.sh:
This basically puts the data into the datastore that is referenced by
the dev_appserver.  To make use of this you will need to run the local
instance and make sure that you are using the lines for large data
operations.  This will ensure that you don't have to do this twice.

server/script_config.sh:
When writing the scripts for downloading and uploading data there were a
lot of shared values and configuring things was becoming confusing.
Factored out the variables into one file that can be sourced into the
other scripts.

server/update_dev_server.sh:
Changed due to refactoring above.

server/run_server_locally.sh:
Changed due to refactoring above, and added some configuration to limit
the amount of debuging messages displayed and sourcing the data
downloaded from GAE.