Adding utility functions for data archival process.
These files will be used to serialize and compress entities from the
datastore on GAE into a zip file.  Handlers for triggering these
functions will be in the next CL.

server/gspeedometer/helpers/archive_util.py - see file commentary for
details on what functions have been added.

server/gspeedometer/helpers/util.py - added a couple comments about
issues that need to be addressed.  Issues will be added in the github
issue tracker.  One problem is that it is possible to have dangling
references when some data is removed and ConvertToDict will throw a
ReferencePropertyResolveError.  The other problem is that
ConvertFromDict does not seem to be able to deal with unconverting
output from ConvertToDict in some cases.

server/gspeedometer/model.py - added a function that will take a string
and return the class that is represented by that string.  This avoids
evaluation of the string and the possible security issues that brings.